+++ modules/canbus/vehicle/lexus/protocol/horn_cmd_11c.cc
@@ -64,7 +64,7 @@ void Horncmd11c::set_p_ignore_overrides(uint8_t* data, bool ignore_overrides) {
   uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 1, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 1, 1);
 }
 
 Horncmd11c* Horncmd11c::set_enable(bool enable) {
@@ -79,7 +79,7 @@ void Horncmd11c::set_p_enable(uint8_t* data, bool enable) {
   uint8_t x = enable;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 0, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 1);
 }
 
 Horncmd11c* Horncmd11c::set_clear_override(bool clear_override) {
@@ -94,7 +94,7 @@ void Horncmd11c::set_p_clear_override(uint8_t* data, bool clear_override) {
   uint8_t x = clear_override;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 2, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 2, 1);
 }
 
 Horncmd11c* Horncmd11c::set_clear_faults(bool clear_faults) {
@@ -109,7 +109,7 @@ void Horncmd11c::set_p_clear_faults(uint8_t* data, bool clear_faults) {
   uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 3, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 3, 1);
 }
 
 Horncmd11c* Horncmd11c::set_horn_cmd(Horn_cmd_11c::Horn_cmdType horn_cmd) {
@@ -126,7 +126,7 @@ void Horncmd11c::set_p_horn_cmd(uint8_t* data,
   uint8_t x = horn_cmd;
 
   Byte to_set(data + 1);
-  to_set.set_value(x, 0, 1);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 1);
 }
 
 }  // namespace lexus
